Fix Rubocop failures
I didn't realise that there were a bunch of failures since Circle wasn't running on the pull requests. The specs all pass (checked this when I was building the release locally) but there are some Rubocop issues.

I've taken a fairly opinionated approach in solving them, bumping the line length to something saner and disabling a few cops on one of our methods which is long, but I feel fairly sensible in situ.